+++ assets/app/view/game/game_data.rb
@@ -3,13 +3,11 @@
 # backtick_javascript: true
 
 require 'game_manager'
-require 'lib/profile_link'
 
 module View
   module Game
     class GameData < Snabberb::Component
       include GameManager
-      include Lib::ProfileLink
 
       needs :allow_clone, default: true
       needs :allow_delete, default: false
@@ -29,27 +27,6 @@ module View
         children = [render_game_data_buttons]
         children << @json if @show_json
         h(:div, props, children)
-      end
-
-      def format_time(ts)
-        return '' unless ts
-
-        t = Time.at(ts.to_i)
-        t > Time.now - 82_800 ? t.strftime('%T') : t.strftime('%F')
-      end
-
-      def info_row(label, value)
-        h(:p, "#{label}: #{value}")
-      end
-
-      def render_game_info
-        user = @game_data['user']
-        [
-          info_row('Description', @game_data['description']),
-          h(:p, ['Host: ', user ? profile_link(user['id'], user['name']) : '']),
-          info_row('Created', format_time(@game_data['created_at'])),
-          info_row('Last Updated', format_time(@game_data['updated_at'])),
-        ]
       end
 
       def render_game_data_buttons
@@ -100,7 +77,7 @@ module View
 
         buttons.concat(render_random_seed)
 
-        h('div.margined', [h(:h3, 'Game Data')] + render_game_info + [h(:div, { style: { marginTop: '1rem' } }, buttons)])
+        h('div.margined', [h(:h3, 'Game Data'), h(:div, { style: { marginTop: '1rem' } }, buttons)])
       end
 
       def render_random_seed
